[MOVE/#111] 상세정보 관련 파일 이동
