--- app/src/main/java/com/napzak/market/data/detailpage/repositoryimpl/DetailPageRepositoryImpl.kt
@@ -1,16 +1,0 @@
-package com.napzak.market.data.detailpage.repositoryimpl
-
-import com.napzak.market.data.detailpage.datasource.DetailPageDataSource
-import com.napzak.market.data.detailpage.mapper.toDomain
-import com.napzak.market.domain.detailpage.model.ProductDetail
-import com.napzak.market.domain.detailpage.repository.DetailPageRepository
-import javax.inject.Inject
-
-class DetailPageRepositoryImpl @Inject constructor(
-    private val dataSource: DetailPageDataSource
-) : DetailPageRepository {
-    override suspend fun getProductDetail(productId: Long): Result<ProductDetail> = runCatching {
-        val response = dataSource.fetchProductDetail(productId)
-        response.data.toDomain()
-    }
-}
