+++ app/src/main/java/com/napzak/market/data/product/service/DetailProductService.kt
@@ -1,11 +1,11 @@
-package com.napzak.market.data.detailpage.service
+package com.napzak.market.data.product.service
 
 import com.napzak.market.core.network.BaseResponse
-import com.napzak.market.data.detailpage.dto.ProductDetailResponse
+import com.napzak.market.data.product.dto.response.ProductDetailResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
 
-interface DetailPageService {
+interface DetailProductService {
     @GET("products/{productId}")
     suspend fun getProductDetail(
         @Path("productId") productId: Long,
