--- app/src/main/java/com/napzak/market/domain/detailpage/repository/DetailPageRepository.kt
@@ -1,7 +1,0 @@
-package com.napzak.market.domain.detailpage.repository
-
-import com.napzak.market.domain.detailpage.model.ProductDetail
-
-interface DetailPageRepository {
-    suspend fun getProductDetail(productId: Long): Result<ProductDetail>
-}
