+++ app/src/main/java/com/napzak/market/domain/product/repository/DetailProductRepository.kt
@@ -1,0 +1,7 @@
+package com.napzak.market.domain.product.repository
+
+import com.napzak.market.domain.product.model.ProductDetail
+
+interface DetailProductRepository {
+    suspend fun getProductDetail(productId: Long): Result<ProductDetail>
+}
